+++ src/frontend/jest.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   moduleFileExtensions: ['ts', 'tsx', 'js', 'css'],
   setupFilesAfterEnv: ['./testSetup.ts'],
-  testMatch: [__dirname + '/**/*.spec.+(ts|tsx|js)'],
+  testMatch: [__dirname + '/js/**/*.spec.+(ts|tsx|js)'],
   testURL: 'https://localhost',
 };